pyproject.toml
[tool.poetry]
name = "pysdbapi"
version = "0.5.0"
description = "pySDBAPI - simple database API. Use a couple of lines to make a request"
license = "MIT"
maintainers = ["Roman <axeman.ofic@gmail.com>"]
authors = ["Roman <axeman.ofic@gmail.com>"]
repository = "https://github.com/axemanofic/pysdbapi"
keywords = ["python", "database", "prettytable"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
]
readme = "README.md"
packages = [{include = "pysdbapi"}]

[tool.poetry.dependencies]
python = "^3.10"
prettytable = {version = "^3.5.0", optional = true}
pyodbc = {version = "^4.0.34", optional = true}
psycopg2 = {version = "^2.9.5", optional = true}
mysql-connector-python ={version = "^8.0.31", optional = true}


[tool.poetry.extras]
pretty = ["prettytable"]
msql = ["mysql-connector"]
mssql = ["pyodbc"]
psql = ["psycopg2"]


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
pylint = "^2.15.4"
mypy = "^0.982"
isort = "^5.10.1"
pytest = "^7.1.3"
pytest-html = "^3.2.0"
lxml = "^4.9.1"
pylint-json2html = "^0.4.0"
coverage = "^6.5.0"
allure-pytest = "^2.11.1"



[tool.poetry.group.doc.dependencies]
mkdocs-material = "^8.5.10"
mkdocs-static-i18n = "^0.50"
mkdocs-redirects = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.isort]
profile = "black"
